Reorganize oxidd-dump features
